Added flights search from database
