+++ src/models/airport.js
@@ -1,22 +1,14 @@
-var Mongoose = require('mongoose');
+﻿var Mongoose = require('mongoose');
 var Schema = Mongoose.Schema;
 
-var flightsSchema = new Schema({ any: {},  //takes raw qpx response and stores
-    updated_at: Date
+var airportSchema = new Schema({
+    code : { type: String, required: true, trim: true },
+    name : { type: String, required: true, trim: true },
+    city : { type: String, required: true, trim: true },
+    state : { type: String, required: true, trim: true },
+    country : { type: String, required: true, trim: true }
 });
 
-flightsSchema.pre('save', function (next) {  //check if data has previously been cached, if yes, then do not create new document
-    // get the current date
-    var self = this;
-  
-});
+var airport = Mongoose.model('airport', airportSchema);
 
-userSchema.pre('update', function (next) {  //update date when new values are cached
-    // get the current date
-    var self = this;
-});
-
-var flights = Mongoose.model('flights', flightsSchema);
-
-exports.flights  = flights;
-	
+exports.airport = airport;
